refactor: extract renderSuccessButton
